+++ packages/node/src/core/evm/handlers/process-transactions.ts
@@ -1,8 +1,8 @@
 import { getGasPrice } from '../gas-prices';
+import * as fulfillments from '../fulfillments';
 import * as logger from '../../logger';
 import * as nonces from '../../requests/nonces';
 import * as state from '../../providers/state';
-import * as transactions from '../transactions';
 import * as utils from '../utils';
 import { EVMProviderState, ProviderState } from '../../../types';
 
@@ -38,7 +38,7 @@ export async function processTransactions(initialState: ProviderState<EVMProvide
   // =================================================================
   // STEP 3: Submit transactions for each wallet
   // =================================================================
-  const receipts = await transactions.submit(state2);
+  const receipts = await fulfillments.submit(state2);
   const successfulReceipts = receipts.filter((receipt) => !!receipt.data);
   successfulReceipts.forEach((receipt) => {
     logger.info(`Transaction:${receipt.data!.hash} submitted for Request:${receipt.id}`, baseLogOptions);
